# apps/cloud-api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TelaFlow Cloud API",
    "description": "API da TelaFlow Cloud \u2014 fase inicial: governan\u00e7a e integra\u00e7\u00e3o, sem persist\u00eancia completa.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/events/{event_id}/media-requirements": {
      "get": {
        "tags": [
          "media-requirements"
        ],
        "summary": "List Media Requirements",
        "operationId": "list_media_requirements_events__event_id__media_requirements_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Media Requirements Events  Event Id  Media Requirements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "media-requirements"
        ],
        "summary": "Create Media Requirement",
        "operationId": "create_media_requirement_events__event_id__media_requirements_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaRequirementCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Media Requirement Events  Event Id  Media Requirements Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events/{event_id}/media-requirements/{media_id}": {
      "get": {
        "tags": [
          "media-requirements"
        ],
        "summary": "Get Media Requirement",
        "operationId": "get_media_requirement_events__event_id__media_requirements__media_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Media Requirement Events  Event Id  Media Requirements  Media Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "media-requirements"
        ],
        "summary": "Update Media Requirement",
        "operationId": "update_media_requirement_events__event_id__media_requirements__media_id__patch",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaRequirementUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Media Requirement Events  Event Id  Media Requirements  Media Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "media-requirements"
        ],
        "summary": "Delete Media Requirement",
        "operationId": "delete_media_requirement_events__event_id__media_requirements__media_id__delete",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Media Requirement Events  Event Id  Media Requirements  Media Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MediaRequirementCreate": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Label"
          },
          "media_type": {
            "type": "string",
            "enum": [
              "video",
              "image",
              "audio",
              "other"
            ],
            "title": "Media Type"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": false
          },
          "scene_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Id"
          },
          "allowed_extensions_hint": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Extensions Hint"
          }
        },
        "type": "object",
        "required": [
          "label",
          "media_type"
        ],
        "title": "MediaRequirementCreate",
        "description": "Cria\u00e7\u00e3o \u2014 `media_id` gerado no servidor."
      },
      "MediaRequirementUpdate": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "media_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "video",
                  "image",
                  "audio",
                  "other"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Type"
          },
          "required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required"
          },
          "scene_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Id"
          },
          "allowed_extensions_hint": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Extensions Hint"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "MediaRequirementUpdate",
        "description": "Atualiza\u00e7\u00e3o parcial (PATCH)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}